Changed numpy.float -> numpy.float32 and added next() call to iter

+++ datasets/box_detect.py
@@ -657,7 +657,7 @@ class BoxDetectDataset(torch.utils.data.Dataset):
         #cluster_centers=means
         dH=600
         dW=3000
-        draw = np.zeros([dH,dW,3],dtype=np.float)
+        draw = np.zeros([dH,dW,3],dtype=np.float32)
         toWrite = []
         final_k=k
         for ki in range(k):
